Modify Improperly Parsed Addresses

Difficult to parse addresses, particularly those that have alphanumeric house ranges as well as units associated with them, may be mis-parsed by postmastr. This function can be used to manually fix mis-parsed address data.

pm_mutate(.data, uid, address, house, house_range, house_frac, house_suf,
    pre_dir, street, street_suf, suf_dir, unit_type, unit_num,
    city, state, postal, locale = "us")

Arguments

.data
A postmastr object created with pm_prep

uid
A pm.uid value to edit.

address
Optional; the new value for the pm.address variable.

house
Optional; the new value for the pm.house variable.

house_range
Optional; the new value (or vector) for the pm.houseRange variable.

house_frac
Optional; the new value for the pm.houseFrac variable.

house_suf
Optional; the new value for the pm.houseSuf variable.

pre_dir
Optional; the new value for the pm.preDir variable.

street
Optional; the new value for the pm.street variable.

street_suf
Optional; the new value for the pm.streetSuf variable.

suf_dir
Optional; the new value for the pm.sufDir variable.

unit_type
Optional; the new value for the pm.unitType variable.

unit_num
Optional; the new value (or vector) for the pm.unitNum variable.

city
Optional; the new value for the pm.city variable.

state
Optional; the new value for the pm.state variable.

postal
Optional; the new value for the pm.postal variable.

locale
A string indicating the country these data represent; the only current option is "us" but this is included to facilitate future expansion.