Parse Street Address

A wrapper around the parse functions that can be used to shorten all of postmastr's code down to a single function call once dictionaries have been created and tested against the data.

pm_parse(.data, input, var, output, newVar, ordinal = TRUE, add_commas = FALSE,
    keep_parsed = "no", keep_ids = FALSE, dirDict, streetDict, suffixDict, unitDict,
    cityDict, stateDict, locale = "us")

Arguments

.data
A source data set to be parsed

input
Describes the format of the source address. One of either "full" or "short". A short address contains, at the most, a house number, street directionals, a street name, a street suffix, and a unit type and number. A full address contains all of the selements of a short address as well as a, at the most, a city, state, and postal code.

var
A character variable containing address data to be parsed

output
Describes the format of the output address. One of either "full" or "short". A short address contains, at the most, a house number, street directionals, a street name, a street suffix, and a unit type and number. A full address contains all of the selements of a short address as well as a, at the most, a city, state, and postal code.

newVar
Name of new variable to store rebuilt address in.

ordinal
A logical scalar; if TRUE, street names that contain numeric words values (i.e. "Second") will be converted and standardized to ordinal values (i.e. "2nd"). The default is TRUE because it returns much more compact clean addresses (i.e. "168th St" as opposed to "One Hundred Sixty Eigth St").

add_commas
A logical scalar; if TRUE, a comma is added both before and after the city name in rebuild addresses. If FALSE (default), no punctuation is added.

keep_parsed
Logical scalar; if TRUE, all parsed elements will be added to the source data after replacement. Otherwise, if FALSE, only the rebuilt address will be added to the source data (default).

keep_ids
Logical scalar; if TRUE, the identification numbers will be kept in the source data after replacement. Otherwise, if FALSE, they will be removed (default).

dirDict
Optional; name of directional dictionary object

streetDict
Optional; name of street dictionary object

suffixDict
Optional; name of street suffix dictionary object

unitDict
Optional; name of unit dictionary object

cityDict
Optional; name of city dictionary object

stateDict
Optional; name of state dictionary object

locale
A string indicating the country these data represent; the only current option is "us" but this is included to facilitate future expansion.